refactor and made unit tests for watchlist feature

+++ server/routes/watchlist.js
@@ -1,49 +1,40 @@
 const express = require('express');
 const router = express.Router();
-const Watchlist = require('../models/watchlist.js');
+const Watchlist = require('../models/watchlist');
 const getUser = require('../helpers/getUser');
+const createWatchlist = require('../helpers/createWatchlist');
 
 /**
  * GET - Watchlist of a user
  */
 router.get('/:id', getUser, async (req, res) => {
-	const idList = res.user.watchlist;
-	const watchlist = await Watchlist.find({_id: {$in: idList}});
-	// TO DO: Error handling
-	res.status(200);
-	res.setHeader('Content-Type', 'application/json');
-	res.json(watchlist);
+	try {
+		const watchlist = await Watchlist.find({userId: res.user._id});
+		res.json(watchlist);
+	} catch (err) {
+		res.status(500).json({ message: err.message });
+	}
 });
 
 /**
  * POST - Add new item to user's Watchlist
  */
 router.post('/:id', getUser, async (req, res) => {
-	// Validate REQ.BODY
-	// Check if watchlist already exist
-	let currentWatchlist = await Watchlist.find({_id: {$in: res.user.watchlist}});
-	const inList = currentWatchlist.filter((item) => item.title === req.body.title);
-	if (inList.length !== 0) {
-		return res.status(403).json('Already in list').send();
+	try {
+		const currentWatchlist = await Watchlist.find({userId: res.user._id});
+		const checkList = currentWatchlist.filter((item) => item.title === req.body.title);
+		if (checkList.length !== 0) {
+			return res.status(403).json({ message: 'Already in list' });
+		}
+		const watchlist = await createWatchlist(res.user, req.body);
+		res.status(201).json(watchlist);
 	}
-	const watchlist = new Watchlist({
-		type: req.body.type,
-		title: req.body.title,
-		path: req.body.path
-	});
-	await watchlist.save();
-
-	// Update User's list of watchlist id's
-	const userWatchlistIds = await Watchlist.find({_id: {$in: res.user.watchlist}});
-	let newWatchlistIds = [];
-	userWatchlistIds.forEach(function (item) {
-		newWatchlistIds.push(item._id);
-	});
-	newWatchlistIds.push(watchlist._id);
-	res.user.watchlist = newWatchlistIds;
-	await res.user.save();
-
-	res.status(201).json(watchlist);
+	catch (err) {
+		if (err.message == 'Bad formatting') {
+			return res.status(400).json({ message: err.message });
+		}
+		res.status(500).json({ message: err.message });
+	}
 });
 
 /**
@@ -51,14 +42,19 @@ router.post('/:id', getUser, async (req, res) => {
  */
 router.delete('/:id/:wid', getUser, async (req, res) => {
 	// Validate REQ.BODY
-	const wid = req.params.wid;
-	await Watchlist.findByIdAndRemove(wid);
-	console.log('Deleted WL');
-
-	const newWatchlistIds = res.user.watchlist.filter(function(id) { return id.toString() !== wid;});
-	res.user.watchlist = newWatchlistIds;
-	await res.user.save();
-	res.status(201).json('Successful');
+	try {
+		const userWatchlist = await Watchlist.find({userId: res.user._id});
+		const wid = req.params.wid;
+		console.log(userWatchlist);
+		if (!userWatchlist.some(item => item._id.toString() === wid)){
+			return res.status(404).json({ message: 'This user does not have this watchlist item'});
+		}
+		await Watchlist.findByIdAndRemove(wid);
+		res.status(201).json({ message: 'Successfully Deleted' });
+	}
+	catch (err) {
+		res.status(500).json({ message: err.message });
+	}
 });
